Proces review feedback.
