+++ EduVPN/Resources/nl.lproj/Localizable.strings
@@ -94,7 +94,7 @@
 "VPN certificate status is unknown." = "VPN certificaat status is onbekend.";
 
 /* User notification */
-"VPN certificate is expiring and needs to be refreshed" = "VPN certificaat verloopt en moet verlengd worden";
+"VPN certificate is expiring and needs to be refreshed" = "VPN-certificaat verloopt en moet vernieuwd worden";
 "The certificate for instance %@ is set to expire %@." = "Het certificaat voor instantie %@ verloopt op %@.";
 
 /* No comment provided by engineer. */
